Fix: Backend stock decremento y frontend venta flow

+++ src/main/java/com/lucasxyz/gestioncancha/Repositories/CajaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lucasxyz.gestioncancha.Entities.Caja;
 
-public interface CajaRepository extends JpaRepository<Caja, Integer> {}
+public interface CajaRepository extends JpaRepository<Caja, Long> {}
